+++ source/renderer/app/components/staking/stake-pools/StakePoolsList.js
@@ -57,9 +57,7 @@ export class StakePoolsList extends Component<Props, State> {
     window.addEventListener('resize', this.handleResize);
   }
 
-  state = {
-    ...initialState,
-  };
+  state = { ...initialState };
 
   // We need to track the mounted state in order to avoid calling
   // setState promise handling code after the component was already unmounted:
@@ -136,7 +134,7 @@ export class StakePoolsList extends Component<Props, State> {
 
     return (
       <div className={componentClasses}>
-        {stakePoolsList.map((stakePool) => {
+        {stakePoolsList.map(stakePool => {
           const isHighlighted = this.getIsHighlighted(stakePool.id);
           const isSelected = selectedPoolId && stakePool.id === selectedPoolId;