+++ source/renderer/app/components/staking/widgets/TooltipPool.js
@@ -1,7 +1,12 @@
 // @flow
 import React, { Component } from 'react';
 import { observer } from 'mobx-react';
-import { defineMessages, intlShape, FormattedMessage } from 'react-intl';
+import {
+  defineMessages,
+  intlShape,
+  FormattedMessage,
+  FormattedHTMLMessage,
+} from 'react-intl';
 import { Button } from 'react-polymorph/lib/components/Button';
 import { Tooltip } from 'react-polymorph/lib/components/Tooltip';
 import { TooltipSkin } from 'react-polymorph/lib/skins/simple/TooltipSkin';
@@ -26,12 +31,10 @@ import copyCheckmarkIcon from '../../../assets/images/check-w.inline.svg';
 import { getColorFromRange, getSaturationColor } from '../../../utils/colors';
 import {
   formattedWalletAmount,
-  shortNumber,
-  formattedLovelaceToAmount,
+  toFixedUserFormat,
 } from '../../../utils/formatters';
 import { rangeMap } from '../../../utils/numbers';
 import { ellipsis } from '../../../utils/strings';
-import globalMessages from '../../../i18n/global-messages';
 import { STAKE_POOL_ID_COPY_FEEDBACK } from '../../../config/timingConfig';
 import {
   THUMBNAIL_HEIGHT,
@@ -58,7 +61,7 @@ const messages = defineMessages({
   rankingTooltip: {
     id: 'staking.stakePools.tooltip.rankingTooltip',
     defaultMessage:
-      '!!!A hierarchical ranking based on the potential rewards you will earn if you delegate the intended amount of stake to this pool, assuming that it reaches saturation.',
+      '!!!<p>A hierarchical ranking based on the potential rewards you will earn if you delegate the intended amount of stake to this pool, assuming that it reaches saturation.</p><p>*Stake pools with the potential rewards estimated at zero have the same ranking. Please set the stake slider to a higher value for more pools to get potential rewards estimated at more than zero.</p>',
     description: '"Rank" tooltip for the Stake Pools Tooltip page.',
   },
   relativeStake: {
@@ -288,9 +291,11 @@ export default class TooltipPool extends Component<Props, State> {
       arrowBottom,
       arrowLeft
     );
-    const colorBandStyle = {
-      background: color,
-    };
+    const colorBandStyle = !this.isGreyColor
+      ? {
+          background: color,
+        }
+      : {};
 
     this.setState({
       componentStyle,
@@ -445,12 +450,16 @@ export default class TooltipPool extends Component<Props, State> {
         top,
         bottom,
       };
-    if (tooltipPosition === 'bottom')
-      return {
+    if (tooltipPosition === 'bottom') {
+      const borderStyle = !this.isGreyColor && {
         borderBottomColor: this.props.color,
+      };
+      return {
+        ...borderStyle,
         left,
         top,
       };
+    }
     return {
       right,
       top,
@@ -472,6 +481,10 @@ export default class TooltipPool extends Component<Props, State> {
     this.setState({ idCopyFeedback: false });
   };
 
+  get isGreyColor() {
+    return !IS_RANKING_DATA_AVAILABLE;
+  }
+
   renderDescriptionFields = () => {
     const { isIncentivizedTestnet } = global;
     const { intl } = this.context;
@@ -480,7 +493,7 @@ export default class TooltipPool extends Component<Props, State> {
       ranking,
       relativeStake,
       producedBlocks,
-      nonMyopicMemberRewards,
+      potentialRewards,
       cost,
       profitMargin,
       saturation,
@@ -492,6 +505,7 @@ export default class TooltipPool extends Component<Props, State> {
       styles.saturationBar,
       styles[getSaturationColor(saturation)],
     ]);
+
     const fields = [
       {
         key: 'saturation',
@@ -505,7 +519,7 @@ export default class TooltipPool extends Component<Props, State> {
                   }}
                 />
               </span>
-              {`${parseFloat(saturation).toFixed(2)}%`}
+              {`${toFixedUserFormat(saturation, 2)}%`}
             </span>
           </div>
         ),
@@ -514,7 +528,7 @@ export default class TooltipPool extends Component<Props, State> {
         key: 'ranking',
         value: (
           <div className={styles.ranking}>
-            {IS_RANKING_DATA_AVAILABLE && nonMyopicMemberRewards ? (
+            {IS_RANKING_DATA_AVAILABLE ? (
               <span
                 style={{
                   background: getColorFromRange(ranking, {
@@ -524,7 +538,14 @@ export default class TooltipPool extends Component<Props, State> {
                   }),
                 }}
               >
-                {ranking}
+                {!potentialRewards.isZero() ? (
+                  ranking
+                ) : (
+                  <>
+                    {numberOfRankedStakePools + 1}
+                    <span className={styles.asterisk}>*</span>
+                  </>
+                )}
               </span>
             ) : (
               <div className={styles.noDataDash}>
@@ -554,9 +575,10 @@ export default class TooltipPool extends Component<Props, State> {
         key: 'relativeStake',
         value: (
           <div className={styles.defaultColor}>
-            <span className={styles.defaultColorContent}>{`${parseFloat(
-              relativeStake
-            ).toFixed(2)}%`}</span>
+            <span className={styles.defaultColorContent}>{`${toFixedUserFormat(
+              relativeStake,
+              2
+            )}%`}</span>
           </div>
         ),
       },
@@ -572,7 +594,7 @@ export default class TooltipPool extends Component<Props, State> {
                 }),
               }}
             >
-              {`${parseFloat(profitMargin).toFixed(2)}%`}
+              {`${toFixedUserFormat(profitMargin, 2)}%`}
             </span>
           </div>
         ),
@@ -609,7 +631,7 @@ export default class TooltipPool extends Component<Props, State> {
         value: (
           <div className={styles.defaultColor}>
             <span className={styles.defaultColorContent}>
-              {shortNumber(producedBlocks)}
+              {toFixedUserFormat(producedBlocks, 0)}
             </span>
           </div>
         ),
@@ -618,16 +640,9 @@ export default class TooltipPool extends Component<Props, State> {
         key: 'potentialRewards',
         value: (
           <div className={styles.defaultColor}>
-            {nonMyopicMemberRewards ? (
-              <span className={styles.defaultColorContent}>
-                {shortNumber(formattedLovelaceToAmount(nonMyopicMemberRewards))}{' '}
-                {intl.formatMessage(globalMessages.unitAda)}
-              </span>
-            ) : (
-              <div className={styles.noDataDash}>
-                <SVGInline svg={noDataDashSmallImage} />
-              </div>
-            )}
+            <span className={styles.defaultColorContent}>
+              {formattedWalletAmount(potentialRewards)}
+            </span>
           </div>
         ),
       },
@@ -641,7 +656,13 @@ export default class TooltipPool extends Component<Props, State> {
               <Tooltip
                 key={field.key}
                 skin={TooltipSkin}
-                tip={intl.formatMessage(messages[`${field.key}Tooltip`])}
+                tip={
+                  <div className={styles.tooltipWithHTMLContent}>
+                    <FormattedHTMLMessage
+                      {...messages[`${field.key}Tooltip`]}
+                    />
+                  </div>
+                }
               >
                 <div className={styles.labelContainer}>
                   <div className={styles.fieldLabel}>
@@ -688,15 +709,7 @@ export default class TooltipPool extends Component<Props, State> {
       idCopyFeedback,
     } = this.state;
 
-    const {
-      id,
-      name,
-      description,
-      ticker,
-      homepage,
-      nonMyopicMemberRewards,
-      retiring,
-    } = stakePool;
+    const { id, name, description, ticker, homepage, retiring } = stakePool;
 
     const componentClassnames = classnames([
       styles.component,
@@ -706,22 +719,23 @@ export default class TooltipPool extends Component<Props, State> {
     const arrowClassnames = classnames([
       styles.arrow,
       styles[`tooltipPosition${capitalize(tooltipPosition)}`],
+      this.isGreyColor ? styles.greyArrow : null,
     ]);
 
     const retirementFromNow = retiring
-      ? moment(retiring).locale(intl.locale).fromNow(true)
+      ? moment(retiring)
+          .locale(intl.locale)
+          .fromNow(true)
       : '';
 
     const idCopyIcon = idCopyFeedback ? copyCheckmarkIcon : copyIcon;
-    const hoverContentStyles = classnames([
+    const hoverContentClassnames = classnames([
       styles.hoverContent,
       idCopyFeedback ? styles.checkIcon : styles.copyIcon,
     ]);
-    const colorBandStyles = classnames([
+    const colorBandClassnames = classnames([
       styles.colorBand,
-      IS_RANKING_DATA_AVAILABLE && nonMyopicMemberRewards
-        ? null
-        : styles.greyColorBand,
+      this.isGreyColor ? styles.greyColorBand : null,
     ]);
 
     return (
@@ -732,11 +746,7 @@ export default class TooltipPool extends Component<Props, State> {
         aria-hidden
         style={componentStyle}
       >
-        {IS_RANKING_DATA_AVAILABLE && nonMyopicMemberRewards ? (
-          <div className={colorBandStyles} style={colorBandStyle} />
-        ) : (
-          <div className={colorBandStyles} />
-        )}
+        <div className={colorBandClassnames} style={colorBandStyle} />
         <div className={arrowClassnames} style={arrowStyle} />
         <div className={styles.container}>
           <h3 className={styles.name}>{name}</h3>
@@ -757,7 +767,7 @@ export default class TooltipPool extends Component<Props, State> {
             onMouseOut={this.onIdMouseOut}
             onBlur={() => {}}
           >
-            <p className={styles.ellipsisContent}>{ellipsis(id, 20, 20)}</p>
+            <p className={styles.ellipsisContent}>{ellipsis(id, 18, 18)}</p>
             <CopyToClipboard text={id} onCopy={this.onCopyId}>
               <Tooltip
                 className={styles.idTooltip}
@@ -766,7 +776,7 @@ export default class TooltipPool extends Component<Props, State> {
                 skin={TooltipSkin}
                 tip={intl.formatMessage(messages.copyIdTooltipLabel)}
               >
-                <div className={hoverContentStyles}>
+                <div className={hoverContentClassnames}>
                   <p className={styles.hoverContentBackground}>
                     {id} <SVGInline svg={idCopyIcon} />
                   </p>