+++ source/renderer/app/themes/daedalus/yellow.js
@@ -312,7 +312,7 @@ export const YELLOW_THEME_OUTPUT = {
     '--theme-loading-status-icons-off-color': '#ea4c5b',
     '--theme-loading-status-icons-unloaded-loading-color': '#2d2d2d',
     '--theme-loading-status-icons-unloaded-syncing-color': '#2d2d2d',
-    '--theme-loading-status-icons-tooltip-color': '#2d2d2d',
+    '--theme-loading-status-icons-tooltip-color': 'var(--rp-pop-over-bg-color)',
     '--theme-loading-spinner-color': '#2d2d2d',
     '--theme-loading-spinner-medium-color': '#2d2d2d',
   },
@@ -753,7 +753,7 @@ export const YELLOW_THEME_OUTPUT = {
     '--theme-staking-stake-pool-background-color': '#ffffff',
     '--theme-staking-stake-pool-border-color': '#e1dac6',
     '--theme-staking-stake-pool-glow-color': 'rgba(45, 45, 45, 0.14)',
-    '--theme-staking-stake-pool-grey-color': 'rgba(45, 45, 45, 0.5)',
+    '--theme-staking-stake-pool-grey-color': '#b2b3b6',
     '--theme-staking-stake-pool-grey-bg-color': 'rgba(45, 45, 45, 0.1)',
     '--theme-staking-stake-pools-search-icon-color': '#2d2d2d',
     '--theme-staking-stake-pool-saturation-background-color':
@@ -1054,6 +1054,10 @@ export const YELLOW_THEME_OUTPUT = {
     '--theme-news-overlay-update-button-hover-text-color': '#fafbfc',
     '--theme-news-overlay-update-button-background-color-hover': '#2d2d2d',
     '--theme-news-overlay-update-button-border-color': '#2d2d2d',
+  },
+  rpPopOver: {
+    '--rp-pop-over-bg-color': 'rgba(45,45,45,0.9)',
+    '--rp-pop-over-text-color': 'white',
   },
 };